+++ backend/garpix_page/models/custombasepage.py
@@ -1,0 +1,20 @@
+import logging
+
+from garpix_page.models import BasePage
+from garpix_utils.string import get_random_string
+
+
+class CustomBasePage(BasePage):
+
+    class Meta:
+        abstract = True
+
+    def clean(self):
+        self._clean_slug()
+        super(CustomBasePage, self).clean()
+
+    def _clean_slug(self):
+        if self.__class__.objects.filter(slug=self.slug).exclude(pk=self.pk).exists():
+            self.slug = f'{self.slug}{get_random_string(size=4)}'
+            message = f'Обнаружен дубль УРЛ для {self.title}. Добавлен символы для уникальности.'
+            logging.warning(message)